refactor: 오류 처리 및 쿼리 관리 개선으로 거래량 전략 최적화(Optimize Volume strategy with improved error handling and query management)
